Refactored and Fixed bug in day 5
Updated code in file "cargo_mover.py" to look cleaner, taking a large amount of similar code in both the part_one and part_two methods and moving it to a separate method. Fixed and issue in day 5 where the program was only capable of running either part_one or part_two method, but now the program can run both without having to reset itself.